change activities models

+++ server/controllers/dto/activity_request_creation_dto.py
@@ -5,4 +5,4 @@ class ActivityRequestCreationDTO(BaseModel):
     employee_id: int
     images: str
     training_information: str
-    adding_kilocalories_count: int
+    adding_kilocalories_count: float
